Let's stop bitmap modularization here for now
